services/mqtt-kafka-bridge/requirements.txt
paho-mqtt==1.6.1
kafka-python==2.0.2
structlog==23.1.0
